feat(partner): two-column detail redesign with a sticky rail (DEC-8 / E1)
Bring the Partner detail page onto the Product-style two-column standard: the
DetailPageHeader over a 1fr/372px grid — ledger / transactions / payments tabs in
the left column, and a sticky right rail carrying the partner's identity, the
balance hero, the «Обороты» totals and «Контакты».

The wide full-width PartnerBalanceCard is reflowed into the new PartnerDetailRail
(a DetailCard stack) and deleted. Balance stays server-computed (hard rule 8);
the «Обороты» subtotals aggregate the served ledger deltas for display.
